pyproject.toml
[tool.poetry]
name = "image-search"
version = "0.1.0"
description = ""
authors = ["yuyi <yuyi.wsy@oceanbase.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
towhee = "^1.1.3"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
opencv-python = "^4.10.0.84"
torchvision = "^0.20.0"
transformers = "^4.45.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tuna"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
